services/ml-pipeline/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4

# Machine Learning
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196  # YOLOv8
mediapipe==0.10.7  # For pose estimation
onnxruntime==1.16.0
tensorflow==2.13.0  # For LSTM models (optional)
scikit-learn==1.3.0  # For clustering and ML utilities

# Image Processing
Pillow==10.1.0
scikit-image==0.22.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Development
pytest==7.4.3
black==23.11.0


